Remove diffx
# Description
This library was unmaintained and, according to scalac-profiler, slowing
our builds by up to 30x. As far as I can tell, this is due to some
(mis?)use of Magnolia's typeclass derivation in Diffx. The test in
quine-core wasn't actually causing the issue (nor were the arbitrary
instances!) but instead, the macro-heavy diffx assertions.

<details>

<summary>Flamegraph</summary>

![flamegraph](https://github.com/thatdot/quine-plus/assets/789922/5b5e4d15-faaa-41b6-b520-2ec48f77ac9d)

</details>

## Type of change
- [x] Bug fix (non-breaking change which fixes an issue)

# How Has This Been Tested?
Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [x] `sbt '; quine-core/test/clean; quine-core/test/compile'`

# Checklist:
- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] New and existing unit tests pass locally with my changes

GitOrigin-RevId: 9c30b68052d16e1f5e576d745248fb51cb6ba48e