Proof of concept query-planner utilizing dependency analysis
# Description
This is a proof of concept update to the query planner for QuinePattern
to demonstrate consuming information from the updated
`quine-query-language` parser to generate more efficient query plans.
Note that, due to the incomplete state of the WIP interpreter, this
change is also only partial.

* Introduces a new dependency graph data structure
* Traverses simple imperative language output from
`quine-query-language` to generate dependency graph
* Compiles dependency graph into a `QuinePattern`

Note that this implementation didn't end up using the actual symbol
table as of yet. Right now, the `SymbolAnalysisPhase` in
`quine-query-language` rewrites all of the identifiers in the program to
make them unique, and this is the essential piece that the dependency
analysis relies on.

This implementation will be revisited in future PRs.

## Type of change
- [X] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

# How Has This Been Tested?

This was tested manually using a custom Quine pattern recipe. Note that
this feature is hidden behind an undocumented feature flag, so the
impact to existing users is minimal. All unit tests for existing
functionality pass as expected.

# Checklist:
- [X] I have performed a self-review of my code
- [X] I have verified my code doesn't add an implementation for
something that already exists
- [X] My changes generate no new warnings
- [X] I have added tests that prove my change is effective and works
- [X] New and existing unit tests pass locally with my changes

GitOrigin-RevId: 8a883387745ad6240428b2f023dd6299603c663f