QU-1958: Adjust product blues to match brand thatDarkBlue, with Icon adjustment
# Description

The 3 thatDot products, Quine, thatDot Streaming Graph, and thatDot
Novelty all should adhere to the branding in this [brand
book](https://drive.google.com/file/d/11Yku8cvJyV4JVXo1Lu4H2WCy7iXffEgH/view?usp=sharing).
This PR aligns the 3 products pages to use the `thatDarkBlue` dark blue
branding.

This PR also updates the thatDot Streaming Graph wordmark icon to match
the branding, which uses a 2-tone white PNG. White for the main thatDot
text, and a grey dot in the icon (this change is pictured below).

Lastly, this PR updates the TSG Docs Page's Icon to use the thatDot "t",
instead of the Quine "Q" (pictured below)

## Type of change

Delete options that are not relevant, add if necessary
- [x] UI Update
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation update
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (rebased to a single commit to keep separate from other
changes)
- [ ] This change requires a documentation update

# How Has This Been Tested?

Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [x] I ran `quine/run` and inspected the color of the `nav-bar`
rendered on the main route, and observed the Quine "Q" as the docs icon.
- [x] I ran `quine-enterprise/run` and inspected the color of the
`nav-bar` rendered on the main route, and observed the thatDot "t" as
the docs icon.
- [x] I ran `novelty/run` and inspected the color rendered on the
`/explorer` route (containing the thatDot icon and the text `Novelty`),
and observed the still remaining thatDot "t" in the docs route.

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] New and existing unit tests pass locally with my changes
- [x] Any changes required of upstream modules have been updated and
have explicit tests in this PR
- [x] Any downstream modules depending on this change have an
integration plan described in the description

---

## Quine

![image](https://github.com/user-attachments/assets/e74da592-ef8f-47cc-8035-6f25a32e9ae5)

## thatDot Streaming Graph (before thatDot icon change)

![image](https://github.com/user-attachments/assets/19e2ed53-afa6-4dfe-8850-75d9a925b7da)

## thatDot Streaming Graph (after thatDot icon change, showing the grey
dot and white wordmark)

![image](https://github.com/user-attachments/assets/6ec0c3ca-6948-473d-99f9-d0ed5164cd77)

## thatDot Novelty

![image](https://github.com/user-attachments/assets/690cc697-07de-4de0-b57d-b3c8ba7c5644)

---

This PR also updates the thatDot Streaming Graph Docs page Icon to use
the thatDot "t"

![image](https://github.com/user-attachments/assets/3e3397fc-0105-4ff8-94e9-cb3659bad75c)

GitOrigin-RevId: 15646665fa40d0e3c3c02df7d74f82026f7522cd