CALL RECURSIVELY subqueries (experimental)
# Description

Adds an extension to the Cypher language for running recursive
subqueries.

## Usage and Syntax

TLDR:

```
CALL RECURSIVELY WITH <variable bindings> UNTIL (<done condition>) {
   <inner query>
}
```

The inner query will repeatedly be given rows with the columns specified
in `<variable bindings>`. The values for these columns come from one of
two places -- either a row previously emitted by the query
(recursively), or the initial value specified in the variable bindings.
For example, if the variable bindings are `0 AS x, 100 AS y`, the inner
query will run first with the single row `x=0, y=100`. It must return
rows containing at least the columns `x` and `y`, which will be fed back
in to another execution of `<inner query>`. This will continue until
either an infinite loop is detected (in which case the query will fail
with an error -- NOT YET IMPLEMENTED), or all rows return `true` when
the `<done condition>` is evaluated.

Put another way: Starting with the definitions in variable bindings, run
`<inner query>`. For each row that `<inner query>` produces, check the
`<done condition>`. If it evaluates to true, return the row. If it
evaluates to false, re-run <inner query> with some of that row's columns
(the ones specified in `<variable bindings>`)

### Walkthrough example

```
CALL RECURSIVELY WITH 0 AS x UNTIL (x > 2) {
  RETURN x+1 AS x, x-3 AS y
} RETURN x, y
```

First invocation: `0 AS x`:
- `RETURN 0+1 AS x, 0-3 AS y` gives us `x=1, y=-3`
- `x > 2` evaluates as `1 > 2` which is false, so we are going to
recurse with the new value of `x=3` and discard the `y=-3` column

Second invocation (recursive): `x=1`
- `RETURN 1+1 AS x, 1-3 AS y` gives us `x=2, y=-2`
- `x > 2` evaluates as `2 > 2` which is false, so we are going to
recurse with the new value of `x=2` and the discard `y=-2` column

Third invocation (recursive): `x=2`
- `RETURN 2+1 AS x, 2-3 AS y` gives us `x=3, y=-1`
- `x > 2` evaluates as `3 > 2` which is true, so we are going to return
the row `x=3, y=-1`

Parent query:
- subquery produced a single row `x=3, y=-1`. We `RETURN x, y` (both
columns)

Final result of the query: 1 row with columns `x=3, y=-1`

Full documentation will be added for
https://thatdot.atlassian.net/browse/QU-1952

## Runtime

The engine takes the form of a new `Query`, `RecursiveSubQuery`. The
semantics of this new query are to repeatedly re-run the `subQuery`,
passing in the columns from `variables` from either the parent scope
(the first time) or from the last run's scope (on subsequent
invocations). Any time a row would be recursed on, the `doneExpression`
is evaluated. If it returns `true`, the recursion stops and the row is
returned. If it returns `false`, the row is fed back in to the subquery
(recursively!)

This depends on openCypher parser-rewriter changes introduced in
https://github.com/thatdot/front-end/pull/5. Once that is merged and
published, the local resolver configuration can be removed from this PR

## Type of change

- [x] New feature (non-breaking change which adds functionality)
- [x] This change requires a documentation update

# How Has This Been Tested?

Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [x] Added new unit tests

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists (Note: this is definitely the only
implementation of full recursion in Cypher, though the repeated graph
query motivating this apparently is solvable with a vanilla Cypher
feature called [quantified path
patterns](https://neo4j.com/docs/cypher-manual/current/patterns/variable-length-patterns/#quantified-path-patterns)
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] ~I have made corresponding changes to the documentation~ Note:
intentionally deferred until after a product discussion
- [x] I have added tests that prove my change is effective and works
- [x] New and existing unit tests pass locally with my changes
- [x] Any changes required of upstream modules have been updated and
have explicit tests in this PR

GitOrigin-RevId: d1d1a70962d004e73d68a467fc2249327e3052a1