Add and fix additional CALL RECURSIVELY tests
# Description

While drafting a CALL RECURSIVELY query for a client, I found that
variable initializers are compiled in the wrong scope. With @bkyrlach,
we corrected the issue so variable initializers rvalues can refer to
values in the parent scope, but the lvalues are declared in the child
(subquery) scope. Also found a couple of additional edge cases that are
deferred until QU-1947 is to be picked back up

## Type of change

Delete options that are not relevant, add if necessary
- [x] Bug fix (non-breaking change which fixes an issue)

# How Has This Been Tested?

Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [x] New unit test added and passes

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have added tests that prove my change is effective and works
- [x] New and existing unit tests pass locally with my changes

GitOrigin-RevId: 0e4ce4bf1af9007ce22b792657369d45af0002da