QU-1855 - PII Logging overhaul
# Description
Implements a method of logging that could be used for our overall
logging strategy.

(Updated 05/16/24)

# Main Idea
- Created a macro for a syntax for creating strings which redacts PII by
default. The syntax of this macro is the same as string interpolation
except that the result of this macro is not a string, but rather a class
`SafeInterpolator` that contains all of the info needed to construct the
string
- In the logging functions, the resulting output string is constructed
using the `SafeInterpolator` and the PII logging settings in the logger
class
- All of the variables/expressions (that are not hard-coded string)
passed into the macro will be replaced with "**REDACTED**" unless the
expressions are wrapped in the `Safe` function or PII logging has been
enabled in the settings.
- A `SafeInterpolator` can also contain an Exception that will be
printed if exception printing is enabled in the PII logging class. This
exception can only be added to an existing `SafeInterpolator` so the
user is forced to provide a description of the error that could replace
the logging of the verbose exception.
- To call the logger with a `log` interpolated string, you need to have
an implicit logConfig in context. This determines if we show unsafe
values, if we show exceptions, and the redaction method.
- A stripped down interpolator called `safe` can be used for strings
where all of the arguments are safe and there are not exceptions to be
possibly printed. Since the output of the `safe` does not rely on the
logConfig settings, the logger can be called using a `safe` interpolated
string without a logConfig.
# Example
```
val A = 123
val B = 456

{
  implicit val logConfig = LogConfig(
    showUnsafe = true,
    showExceptions = false
  )
  logger.warn(log"variable1 $A variable2 is ${Safe(B)}")
  // A and B are printed  because showUnsafe is true
  // Prints: variable1 123 variable2 is 456
}
{
   implicit val logConfig = LogConfig(
    showUnsafe = false,
    showExceptions = false
  )
  logger.warn(log"variable A as unsafe is $A variable b is ${Safe(B)}".withException(new Exception("example")))
  // A is redacted by default because showUnsafe is false, but B is shown because it is marked as Safe
  // Prints: variable A as unsafe is **REDACTED** variable b is 456
  // No exceptions are printed because it was disabled in the settings
}
{
   implicit val logConfig = LogConfig(
    showUnsafe = false,
    redactor = s => s.reverse,
    showExceptions = true)
  )
  logger.warn(log"variable $A variable2 is ${Safe(B)}".withException(new Exception("example")))
  // A is hidden using our super secure redacting function (reverse) and B is shown because it was marked as Safe
  // Prints: variable 321 variable2 is 456
  // Exception("example") is printed
}
{
   implicit val logConfig = LogConfig(
    showUnsafe = false,
    redactor = verySecureRedactorFunction,
    showExceptions = false)
  val birthMonth = "January"
  logger.warn(log"Birth month = $birthMonth variable2 is ${Safe(B)}")
  // Prints: Birth month = Anuaryjay variable2 is 456
}
logger.warn(log"variable1 is $A") //Does not compile because it could not find an implicit LogConfig
logger.warn(safe"variable1 is $A") //Does not compile because A is not safe
logger.warn(safe"variable1 is ${Safe(A)}" //Prints "variable 1 is 123"
```

# Implementation details
- Created a type class `Loggable` that has a function `safe` for
encoding values that we know do not contain PII and a function `unsafe`
that provides the encoding for redacted data
- `unsafe` has an extra parameter of type `string => string` that is the
obfuscation method. This is so we can separate the logic of the encoding
and obfuscation method
- When the logging functions are called, it uses the config to determine
whether to encode the arguments using the `safe` or `unsafe` functions
(remember that even if we call `unsafe` any value wrapped in the Safe
class will be displayed as if we called `safe`)

Co-authored-by: David Masters <david@thatdot.com>
Co-authored-by: Ethan Bell <ethan@thatdot.com>
GitOrigin-RevId: 262f510499d42d62d6dce61fcbf1fe06017126ee