Add CI scripts and GitHub actions for trial version Docker image upload
# Description

The Quipu engagement has recently required the delivery of Helm charts
for thatDot products. As a dependency of those Helm charts, Docker
images for the trial versions should be publicly available to enable
customers to deploy the Helm charts with minimal friction. This PR adds
scripts and CI automation for building and uploading those images with
each release of Streaming Graph and Novelty.

- Add scripts in `ci` to build both Streaming Graph and Novelty trial
docker images
- Add GitHub Actions to build Streaming Graph and Novelty trial docker
images.
- Call these new actions from existing release workflows
  - Streaming Graph Release
  - Novelty Release
  - Quine and Streaming Graph Release

This PR also incidentally fixes a recently discovered bug where the
version of novelty was not correctly being picked up by CI, causing an
error in the generated jar names when uploaded to S3.

## Type of change

- [x] New feature (non-breaking change which adds functionality)
- [x]  Bug fix (non-breaking change which fixes an issue)

# How Has This Been Tested?

Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [x] Scripts have been tested manually

Actions and workflows are currently untested as this is more difficult
when modifying release workflows. Mitigating factor - the changes are
low-risk.

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] New and existing unit tests pass locally with my changes
- [x] Any changes required of upstream modules have been updated and
have explicit tests in this PR
- [x] Any downstream modules depending on this change have an
integration plan described in the description

GitOrigin-RevId: 70a75886094833eb4134312e2c6b9f07489f7732