QU-1930: Fix `standing.wiretap` error message
# Description

Tiny change to fix the error message of standing.wiretap when it cannot
find a standing query. After the change the error message should
correctly report the name of the not-found standing query.

## Type of change

Delete options that are not relevant, add if necessary
- [x] Bug fix (non-breaking change which fixes an issue)

# How Has This Been Tested?

Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [x] Manually test standing.wiretap with incorrect name

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my change is effective and works
- [x] New and existing unit tests pass locally with my changes
- [x] Any changes required of upstream modules have been updated and
have explicit tests in this PR
- [x] Any downstream modules depending on this change have an
integration plan described in the description

GitOrigin-RevId: 82d36b797cf677c72c92435c38037daaf07ffab3