Omit labelsProperty from MVSQ view of properties
# Description

Ad-hoc queries don't actually see a node's `properties` when they
request a node's properties, they see what I'll call a node's "cypher
properties". These are basically identical, except they exclude the
`labelsProperty` (which is what we use to store the list of Cypher
labels a node has). This PR introduces that same mapping rule to the
MVSQ interpreter so the query `MATCH (n) RETURN properties(n)` now
behaves the same as an MVSQ as it does as an ad-hoc query

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)

# How Has This Been Tested?

Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [x] Manually verified with the attached REST requests

[mvsq-allproperties-verif.rest.txt](https://github.com/user-attachments/files/16475846/mvsq-allproperties-verif.rest.txt)

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] New and existing unit tests pass locally with my changes

GitOrigin-RevId: 74463a908c3bb43325552b95fb295e5f286ff3c7