Replace straggling `qid.debug` with `qid.pretty`
# Description

Quite a while ago, we defined `QuineId.pretty` as a way to serialize
QuineIds in a human-readable form. This is also what `qid.debug` was
intended for. There were some differences in the intended audience, but
nothing that practically outweighs the confusion introduced by the same
value appearing in multiple styles in log messages. Begone, `debug`!

## Type of change

Delete options that are not relevant, add if necessary
- [x] Refactor (rebased to a single commit to keep separate from other
changes)

# How Has This Been Tested?

- [x] CI passes

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] New and existing unit tests pass locally with my changes

GitOrigin-RevId: 12945ea19a7419b294c0817ff9ef5b8670787cef