Remove unnecessary unicode (fancy quotes, ellipsis) for consistency and compatibility
# Description

In some cases, this was causing examples to not work. In all cases, it's
inconsistent.

I have left the emdashes, because there is enough visual distinction
between `--` and `—` that someone might notice. I have adjusted `--`
used in markdown files to be `—` instead, for consistency. I also added
spaces before and after. While this is debatable, it makes these easier
to identify in a search (`--` may have special meaning when used
alongside other symbols, as in a markdown table `|--` or section `--`)

There is an argument to be made that these smart quotes, etc. should
remain in docs, but if so, we should use them _consistently_ to avoid
distracting the reader. On a more conceptual level, this is one of the
reasons to write docs in markdown - so that our typesetter can apply
fancy punctuation on our behalf, in a uniform manner, if that's how we
want our content presented.

By the way, these can be disabled on Mac here:
<img width="681" alt="image"
src="https://github.com/user-attachments/assets/d116108f-6341-4c19-afeb-ec13a7ce625e">

Also, found one place where `DistinctId` was being referred to as
`SingleId`. I didn't even remember that name!

## Type of change

Delete options that are not relevant, add if necessary
- [x] Bug fix (non-breaking change which fixes an issue)
- [x] Documentation update

# How Has This Been Tested?

Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [x] CI passes

# Checklist:

- [x] I have performed a self-review of my code

GitOrigin-RevId: 599fd2487cdfbb982038f9f4298121edf93c0b0a