QU-1963: Caught PatternSyntaxExceptions and threw CypherException
# Description

A few regex-based functions could fail to compile the provided regular
expression and throw an exception which amounts to a user error – this
exception should therefore be presented to the user. For the ad-hoc
interpreter, this means throwing a CypherException with an appropriate
error message.

## Type of change

Delete options that are not relevant, add if necessary
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation update
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (rebased to a single commit to keep separate from other
changes)
- [ ] This change requires a documentation update

# How Has This Been Tested?
Added a test to the CypherExpressions tests to make sure it throws the
correct exception

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] I have added tests that prove my change is effective and works
- [x] New and existing unit tests pass locally with my changes
- [x] Any changes required of upstream modules have been updated and
have explicit tests in this PR
- [x] Any downstream modules depending on this change have an
integration plan described in the description

---------

GitOrigin-RevId: b7cdab764b6a6794516d1b7d59f1f09a8a73b150