Fix for QU-1961
# Description

This change hopes to eliminate the CVE introduced with the `CALL
RECURSIVELY` change. I believe the CVE was introduced when a dependency
for `front-end` was changed from being `test` scope during some
troubleshooting.

## Type of change

Delete options that are not relevant, add if necessary
- [X] Bug fix (non-breaking change which fixes an issue)

# How Has This Been Tested?

Checking dependency tracker to see if the CVE is no longer listed.

# Checklist:

This is just a one line dependency update.

GitOrigin-RevId: bafbdb7f677b766bd9fd5aed7e17c2b23f9c177c