QU-2010: Convert `GenerateEnterpriseOpenApi` into `object` from `class`
# Description

This is a bug fix enabling the `main` entrypoint for the
`quine-enterprise-docs` project to be ran when using:
- `makeSite`
- `previewSite`

Since `GenerateEnterpriseOpenApi` is a program entrypoint, it is
required to be `static` which occurs when the program entrypoint which
`extends App` is an `object`, otherwise a `NoSuchMethodException` is
thrown:

```
[error] java.lang.NoSuchMethodException: com.thatdot.quine.docs.GenerateEnterpriseOpenApi.main is not static
```

Converting `GenerateEnterpriseOpenApi` back into an `object` resolves
this exception.

To ensure this doesn't happen again w/ `GenerateEnterpriseOpenApi` since
this is untested, I added a single test to ensure that the project
entrypoint is asserted to be a static object.

## Type of change

Delete options that are not relevant, add if necessary
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation update
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (rebased to a single commit to keep separate from other
changes)
- [ ] This change requires a documentation update

# How Has This Been Tested?

Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [x] I ran `makeSite` and `previewSite` to ensure that the docs can be
generated
- [x] I created `GenerateEnterpriseOpenApiTest` to capture this bug in a
test to ensure that the main is asserted to be a static object.

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have added tests that prove my change is effective and works
- [x] New and existing unit tests pass locally with my changes

GitOrigin-RevId: 469b00e70af91ee44110c0c774eb579cc41a2139