QU-1954: Ensure Novelty Context Dropdown on Plots Page always shows updated list of contexts
# Description

This PR ensures that, in Novelty, the Context Dropdown on the Plots page
always has an updated list of contexts.

The problem is that the plots page will only call out to Novelty for a
list of contexts one single time, the very first time a user lands on
the page.

If a user:
- looks at the plots page first
- and then navigates to the upload page and performs an upload to a new
context
- and then navigates back to the plots page

The new context will not be present in the dropdown.

---

The reason this is happening is because the current router used by our
products uses CSS to hide and reshow pages. It does not mount/unmount
them from the DOM. This means `componentDidMount` will only run one
single time.

This PR adds a bit more logic to `PageWithSideBar`, adding an optional
mount function to the `Tab` type, and having the `PageWithSideBar` run
that function on mount, and also when switching to a Tab.

## Type of change

Delete options that are not relevant, add if necessary
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation update
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (rebased to a single commit to keep separate from other
changes)
- [ ] This change requires a documentation update

# How Has This Been Tested?

Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [x] I ran Novelty and observed the network tab in DevTools to ensure
the gremlin call for all contexts occured each time I switched to the
plots route.
- [x] I ran quine to ensure my messing around with the client-side
router didnt negatively affect rendering there (I tried to keep changes
to `PageWithSideBar` as minimal as possible)

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas

GitOrigin-RevId: 12b8ad496dfa6781feab213ef8408452e213d6ba