QU-2029/serialize retriable failures
# Description

Added a generalized error type that represents exceptions. This creates
three categories of errors:

QuineErrors which are exceptions defined by Quine

ExternalErrors which are exceptions that are defined outside of Quine
(such as in pekko)

GenericError which is a class that represents exceptions that we haven't
defined a generalized error for.

These errors currently just are a wrapper around existing exceptions,
but eventually we would phase out exceptions defined in Quine almost
entirely. Philosophically, this is because we should only throw
exceptions when there is an issue with Quine itself (OOM, null pointer
exception, divide by zero, etc), but we shouldn't throw exceptions due
to expected issues due to user error. For example if the user provides a
recipe file that doesn't exist, has a misconfigured config, or provides
an invalid cypher query, then the error isn't from Quine and we just
need a descriptive message that tells the user what went wrong.

AnyError/BaseError does not extend throwable because this is intended as
a first step in moving away from exceptions whenever possible. As we
transition away from exceptions, the exception fields in each error
would be replaced with just the information required to describe the
error and its cause.

## Type of change

Delete options that are not relevant, add if necessary
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation update
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (rebased to a single commit to keep separate from other
changes)
- [ ] This change requires a documentation update

# How Has This Been Tested?

I ran two tests to check to make sure I am reproducing the bug and that
I have fixed it. First, on main, I changed the CypherMetaType::call
function to

```  def call(args: Vector[Value])(implicit idp: QuineIdProvider, logConfig: LogConfig): Value = {
      throw new ExactlyOnceTimeoutException("fail");
      args match {
        case Vector(v) => Expr.Str(v.typ.pretty)
        case other => throw wrongSignature(other)
      }
    }
```

And ran a numberiteratoringest on host index 0, that should produce a
retryable error using: the query
```
MATCH (n) WHERE id(n) = locIdFrom(1, $that) SET n.prop = 5 SET n.t = meta.type(n.prop) RETURN n
```

And verified that the ingest fails and I saw this error in the logs:
```
2024-08-28 12:54:02,596 WARN [NotFromActor] [test-cluster-pekko.actor.default-dispatcher-18] com.thatdot.quine.app.routes.IngestStreamWithControl - Ingest stream: NumberIteratorIngest(CypherJson(MATCH (n) WHERE id(n) = locIdFrom(1, $that) SET n.prop = 5 SET n.t = meta.type(n.prop) RETURN n,that),0,None,None,16) failed.
```

Then, I ran the same test with the same query on this branch and the
result was that the ingest succeed, but the ingest count was stuck at 16
(which I believe is expected)

The two test above show that this branch fixes the bug, but additionally
I ran the tests again on this branch, first with replacing the
ExactlyOnceTimeoutException with a NullPointerException and verified the
query failed as expected. Then as one final test, I returned the
CypherMetaType::call to normal by removing the exceptions and verified
that the ingest stream ran as expected.

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
~- [ ] I have added tests that prove my change is effective and works~
- [x] New and existing unit tests pass locally with my changes
- [ ] Any changes required of upstream modules have been updated and
have explicit tests in this PR
- [x] Any downstream modules depending on this change have an
integration plan described in the description

GitOrigin-RevId: fec4dd52da3adfc13bee458fe55408a02158216b