Mirror dependencySchemes.sbt to open-source repo to fix CI
# Description

#2895 seems to have introduced conflicting versions of scala-xml to the
open source mirror of the repository (first failure:
https://github.com/thatdot/quine/actions/runs/10411409949/job/28835260195)

We have had issues with this dependency in the past, and @cmeade has
even written a handy fix as a project-level build file. Unfortunately,
that fix was not included in the repo mirroring configuration, so our
open source repo does not benefit from it.

This adds the relevant fix file to our repo mirroring configuration to
fix the open source CI

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)

# Checklist:

- [x] I have performed a self-review of my code

GitOrigin-RevId: 64a34cb3f672dd2cc927a51b0b5e602f6063c892