Initial implementation of QuinePatternRegistry
# Description

This is an initial (rough) implementation of a query plan registry. This
is similar to (but broader in scope) the StandingQueryPattern registry.
This will be expanded in the future to correctly work with the
persistence layer and clustered operations.

## Type of change

Delete options that are not relevant, add if necessary
- [X] New feature (non-breaking change which adds functionality)

# How Has This Been Tested?

This registry component can't yet be validated. Validation will come
when the query loader component is more fleshed out.

# Checklist:

- [X] I have performed a self-review of my code
- [X] I have verified my code doesn't add an implementation for
something that already exists
- [X] New and existing unit tests pass locally with my changes

GitOrigin-RevId: 6661b6f543b666a6fa33cdf71043da92c63a5a68