Handle an exceptional case in MVSQ loading
# Description

When a node receives a NewMultipleValuesState result, we have a log line
that was making the assumption that the result was for an SQ that still
exists. This could throw an exception if the standing query was deleted
by the time the message was received. Luckily, this information was only
used for a log, and so we can simply change the log message if the part
was deleted

## Type of change

Delete options that are not relevant, add if necessary
- [x] Bug fix (non-breaking change which fixes an issue)

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] New and existing unit tests pass locally with my changes

GitOrigin-RevId: 336cd6a20136317f22d11c1534b6f0e82a2d9e8f